raise an error when calling cell_normals before compute_normals()
Right now if you call cell_normals before calling compute_normals(), it will simply return an empty array.  This is incorrect, and is especially confusing because if a user is new to vedo they will have no way of knowing that there's another function they need to call first unless they happen to notice it in the API.

This commit has it throw an error instead, thus preventing it from returning the wrong answer and informing the user how to fix the problem.